new devices
- amg8833 ir matrix camera
- 5x5 led matrix

+++ cmd/dispsim/dispsim_const.go
@@ -1,9 +1,9 @@
 package main
 
-const(
+const (
 	PixelEdgeDimension = 64
 	PixelSpacing       = 8
 
-	DefaultResolutionX = 5
-	DefaultResolutionY = 10
+	DefaultResolutionX = uint(5)
+	DefaultResolutionY = uint(10)
 )
